test: DtoFactory의 jobResponses 메소드를 DummyJob으로 이동
