+++ backend/src/test/java/our/portfolio/devspace/utils/DtoFactory.java
@@ -1,13 +1,8 @@
 package our.portfolio.devspace.utils;
 
-import java.util.ArrayList;
-import java.util.List;
-import our.portfolio.devspace.domain.job.dto.JobResponse;
-import our.portfolio.devspace.domain.job.entity.JobType;
 import our.portfolio.devspace.domain.post.dto.CreatePostRequest;
 import our.portfolio.devspace.domain.profile.dto.CreateProfileRequest;
 import our.portfolio.devspace.domain.profile.dto.SimpleProfileResponse;
-import our.portfolio.devspace.utils.dummy.DummyJob;
 import our.portfolio.devspace.utils.dummy.DummyPost;
 import our.portfolio.devspace.utils.dummy.DummyProfile;
 
@@ -41,18 +36,5 @@ public class DtoFactory {
             .company(profile.getCompany())
             .name(profile.getName())
             .build();
-    }
-
-    public static List<JobResponse> jobResponses(JobType type) {
-        List<JobResponse> jobResponses = new ArrayList<>();
-
-        for (int i = 1; i <= 10; i++) {
-            DummyJob job = new DummyJob(i);
-            job.setType(type);
-            job.setTitle(type.toString() + i);
-            jobResponses.add(new JobResponse(i, job.getTitle()));
-        }
-
-        return jobResponses;
     }
 }
